only delete packages if they exist

+++ packages/nextjs/vercel/set-up-branch-for-test-app-use.sh
@@ -11,13 +11,21 @@ if [ -n "${STASHED_CHANGES}" ]; then
   git stash -u
 fi
 
-# get rid of irrelevant packages to speed up deploy process and commit the result
+# if this hasn't already been done, get rid of irrelevant packages to speed up deploy process and then commit the result
+PACKAGES_DELETED=false
 for package in "angular" "ember" "eslint-config-sdk" "eslint-plugin-sdk" "gatsby" "serverless" "vue" "wasm"; do
-  echo "deleting ${package}"
-  rm -rf packages/${package}
+  if [ -d packages/${package} ]; then
+    echo "deleting ${package}"
+    rm -rf packages/${package}
+    PACKAGES_DELETED=true
+  fi
 done
-git add .
-git commit -m "delete unneeded packages"
+
+if [ $PACKAGES_DELETED = true ]; then
+  echo "committing deletions"
+  git add .
+  git commit -m "delete unneeded packages"
+fi
 
 # restore working directory, if necessary
 if [ -n "${STASHED_CHANGES}" ]; then
